requirements.txt
alembic==1.0.8
Click==7.0
dominate==2.3.5
enum34==1.1.6
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
slugify==0.0.1
SQLAlchemy==1.3.2
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
